fix: change max_change type from uint16_t to float for precision

+++ esphome/components/sofarsolar_inverter/sofarsolar_inverter.cpp
@@ -19,7 +19,7 @@ namespace esphome
 			bool write_set_value = false; // Flag to indicate if the write value is set
 			bool is_queued = false; // Flag to indicate if the register is queued for reading/writing
 			bool is_max_change_flat = false; // Flag to indicate if the register value is a flat change
-			uint16_t max_change = 0; // Maximum allowed change for the register value
+			float max_change = 0; // Maximum allowed change for the register value
 			SofarSolar_RegisterDynamic() : sensor(nullptr), update_interval(0), last_update(0), default_value({}), default_value_set(false), enforce_default_value(false) {}
 		};
